requirements.txt
Django==3.0.8
djangorestframework==3.11.0
Pillow==7.2.0
py3-validate-email==0.2.9
dnspython==1.16.0
psycopg2==2.8.5
drf-yasg==1.17.1
celery==4.4.6
django-celery-beat==2.0.0
djangorestframework-simplejwt==4.4.0
redis==*
